调整工程，去除警告“warning MSB8004: Intermediate 目录未以斜杠结尾。”
git-svn-id: svn://192.168.2.188/qtcms@381 894f9759-fa2a-45cf-9159-0a04a166245e

+++ src/functionlibs/RemotePreview/h264wh.cpp
@@ -44,7 +44,7 @@ int Se(BYTE *pBuff, UINT nLen, UINT &nStartBit)
 
 int UeVal=Ue(pBuff,nLen,nStartBit);
 double k=UeVal;
-int nValue=ceil(k/2);//ceil函数：ceil函数的作用是求不小于给定实数的最小整数。ceil(2)=ceil(1.2)=cei(1.5)=2.00
+int nValue=(int)ceil(k/2);//ceil函数：ceil函数的作用是求不小于给定实数的最小整数。ceil(2)=ceil(1.2)=cei(1.5)=2.00
 if (UeVal % 2==0)
 nValue=-nValue;
      return nValue;
